Moves RockTheVoteRecord class into separate file
